feat: order aggregate factory 구현
- order aggregate 등록을 위한 factory 구현
