+++ src/main/java/dev/practice/order/domain/order/OrderItemSeriesFactory.java
@@ -1,0 +1,9 @@
+package dev.practice.order.domain.order;
+
+import dev.practice.order.domain.order.item.OrderItem;
+
+import java.util.List;
+
+public interface OrderItemSeriesFactory {
+    List<OrderItem> store(Order order, OrderCommand.RegisterOrder requestOrder);
+}
