+++ src/main/java/dev/practice/order/domain/order/OrderServiceImpl.java
@@ -1,6 +1,5 @@
 package dev.practice.order.domain.order;
 
-import dev.practice.order.domain.item.ItemReader;
 import lombok.RequiredArgsConstructor;
 import lombok.extern.slf4j.Slf4j;
 import org.springframework.stereotype.Service;
@@ -11,27 +10,13 @@ import org.springframework.transaction.annotation.Transactional;
 @RequiredArgsConstructor
 public class OrderServiceImpl implements OrderService {
     private final OrderStore orderStore;
-    private final ItemReader itemReader;
+    private final OrderItemSeriesFactory orderItemSeriesFactory;
 
     @Override
     @Transactional
     public String registerOrder(OrderCommand.RegisterOrder requestOrder) {
         Order order = orderStore.store(requestOrder.toEntity());
-        requestOrder.getOrderItemList().forEach(orderItemRequest -> {
-            var item = itemReader.getItemBy(orderItemRequest.getItemToken());
-            var initOrderItem = orderItemRequest.toEntity(order, item);
-            var orderItem = orderStore.store(initOrderItem);
-
-            orderItemRequest.getOrderItemOptionGroupList().forEach(orderItemOptionGroupRequest -> {
-                var initOrderItemOptionGroup = orderItemOptionGroupRequest.toEntity(orderItem);
-                var orderItemOptionGroup = orderStore.store(initOrderItemOptionGroup);
-
-                orderItemOptionGroupRequest.getOrderItemOptionList().forEach(orderItemOptionRequest -> {
-                    var initOrderItemOption = orderItemOptionRequest.toEntity(orderItemOptionGroup);
-                    orderStore.store(initOrderItemOption);
-                });
-            });
-        });
+        orderItemSeriesFactory.store(order, requestOrder);
         return order.getOrderToken();
     }
 }
